refactor: extract spawnEngine with browser strategy
